+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project follows [Semantic Versioning](https://semver.org/spec/v2.0.0.ht
 - Added additional .info entries for Granger analysis, indicating details about the computation.
 - Added additional .info entries for FoooF results, e.g. Gaussian fit parameters.
 - Fix bug #365, plotting supports custom dimords now.
+- Support starting a connectivity analysis from a SpectralData instance (#364).
 
 ### CHANGED
 - Maximal brute force regularization parameter for Granger increased to 1e-1